agregator/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
build==1.2.2
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.8
inflection==0.5.1
multidict==6.1.0
packaging==24.1
psycopg2-binary==2.9.9
pyproject_hooks==1.1.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.11.1
